# astonService/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: User API
  version: 1.0.0
paths:
  /users:
    get:
      operationId: getAllUsers
      summary: Получить всех пользователей
      description: Возвращает список всех пользователей.
      tags:
        - user
      responses:
        '200':
          description: Список пользователей успешно получен.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponse'
    post:
      operationId: createUser
      summary: Создать нового пользователя
      description: Создает нового пользователя с переданными данными.
      tags:
        - user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreateRequest'
      responses:
        '201':
          description: Пользователь успешно создан.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserEvent'
        '400':
          description: Неверные данные для создания пользователя.

  /users/{id}:
    get:
      operationId: getUser
      summary: Получить информацию о пользователе по ID
      description: Возвращает информацию о пользователе с указанным ID.
      tags:
        - user
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
          description: Идентификатор пользователя.
      responses:
        '200':
          description: Пользователь найден.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          description: Пользователь не найден.
    put:
      operationId: updateUser
      summary: Обновить информацию о пользователе
      description: Обновляет информацию о пользователе с указанным ID.
      tags:
        - user
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
          description: Идентификатор пользователя.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateRequest'
      responses:
        '200':
          description: Пользователь успешно обновлен.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserEvent'
        '400':
          description: Неверные данные для обновления пользователя.
        '404':
          description: Пользователь не найден.
    delete:
      operationId: deleteUser
      summary: Удалить пользователя
      description: Удаляет пользователя по указанному ID.
      tags:
        - user
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
          description: Идентификатор пользователя.
      responses:
        '204':
          description: Пользователь успешно удален.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserEvent'
        '404':
          description: Пользователь не найден.

  /users/event:
    post:
      operationId: sendUserEvent
      summary: Отправить событие пользователя в Kafka
      description: Позволяет публиковать событие (создание, обновление, удаление) пользователя
      tags:
        - user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserEvent'
      responses:
        '200':
          description: Событие успешно отправлено
        '400':
          description: Неверные данные события

components:
  schemas:
    UserCreateRequest:
      type: object
      required:
        - name
        - email
        - age
      properties:
        name:
          type: string
          description: Имя пользователя.
          minLength: 2
          maxLength: 50
        email:
          type: string
          format: email
          description: Адрес электронной почты пользователя.
        age:
          type: integer
          description: Возраст пользователя.
          minimum: 1
          maximum: 150

    UserUpdateRequest:
      type: object
      required:
        - name
        - email
      properties:
        id:
          type: integer
          format: int64
          description: Уникальный идентификатор пользователя.
        name:
          type: string
          description: Имя пользователя.
          minLength: 2
          maxLength: 50
        email:
          type: string
          format: email
          description: Адрес электронной почты пользователя.
        age:
          type: integer
          description: Возраст пользователя.
          minimum: 1
          maximum: 150

    UserResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Уникальный идентификатор пользователя.
        name:
          type: string
          description: Имя пользователя.
        email:
          type: string
          format: email
          description: Адрес электронной почты пользователя.
        age:
          type: integer
          description: Возраст пользователя.
        createdAt:
          type: string
          format: date-time
          description: Дата и время создания пользователя.

    UserEvent:
      type: object
      required:
        - email
        - eventType
      properties:
        email:
          type: string
          format: email
          description: Адрес электронной почты пользователя
        eventType:
          type: string
          enum:
            - CREATED
            - UPDATED
            - DELETED
          description: Тип события (создание, обновление или удаление пользователя)